(PUP-1840) Ease testing under different digest_algorithm settings
Without this patch, spec tests for pieces of the code having to do
with checksums would have to be duplicated twice and contain repeated
code.

The patch adds the `using_checksums_describe` method, which can be used
instead of the `describe` method to enclose spec tests where checksums
are involved. This method only defined inside describe blocks where the
`:uses_checksums => true` user metadata is given, to mitigate the risk
of unintended changes to the behavior of spec tests not having to do
with checksums.

`using_checksums_describe` defines a group of examples, each having the
given block as body. Each example happens with a different value set
for `digest_algorithm`. Many `let`s are defined for use inside the
spec test; foremost are `plaintext`, `checksum`, and
`algo`. `plaintext` is some text; `checksum` is the checksum for this
text under the present setting of `digest_algorithm`. `algo` contains
the name of the digest algorithm in use.
unset, `algo` is 'md5'.

Code restructured and simplified by Adrien Thebo <adrien@puppetlabs.com>